+++ CHANGELOG.md
@@ -4,9 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Respect browser's `Accept-Language` header for date time display
+- Add new schedule item setting `Fill With Group Mode`
+  - This setting is only available when a `Collection`, `Multi-Collection` or `Smart Collection` is scheduled with `Duration` or `Multiple` playout modes
+  - For this setting, groups are *only* television shows and artists
+  - The options are:
+    - `None`: no change to scheduling behavior - all groups (shows and artists) will be shuffled/ordered together
+    - `Ordered Groups`: each time this item is scheduled, the entire `Duration` or `Multiple` will be filled with a single group, and the groups will rotate in a fixed order
+    - `Shuffled Groups`: each time this item is scheduled, the entire `Duration` or `Multiple` will be filled with a single group, and the groups will rotate in a shuffled order
+
 ### Fixed
 - Fix error loading path replacements when using MySql
 - Fix tray icon shortcut to open logs folder on Windows
+
+### Changed
+- Upgrade from .NET 7 to .NET 8
 
 ## [0.8.4-beta] - 2023-12-02
 ### Fixed